b-4-get-movies-by-genre: 1) Add @ManyToMany mapping on entities; 2) Add new method into the DefaultMovieService, add test case for it; 3) Refactoring: 3.1) Merged JpaConfig into MovielandApplicationContext; 3.2) Removed hibernate.cfg.xml and add its properties into hibernate.properties; 3.3) Moved persistence.xml file's properties into hibernate.properties;
