--- hibernate.cfg.xml
@@ -1,38 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD//EN"
-        "http://hibernate.sourceforge.net/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-
-    <session-factory>
-
-        <!-- Database connection settings -->
-        <property name="connection.driver_class">org.postgresql.Driver</property>
-        <property name="connection.url">jdbc:postgresql://localhost:5432/movieland-db</property>
-        <property name="connection.username">postgres</property>
-        <property name="connection.password">root</property>
-
-        <!-- SQL dialect -->
-        <property name="dialect">org.hibernate.dialect.PostgreSQL9Dialect</property>
-
-        <!-- JDBC connection pool (use the built-in) -->
-        <property name="connection.pool_size">1</property>
-
-        <!-- Enable Hibernate's automatic session context management -->
-        <property name="current_session_context_class">thread</property>
-
-        <!-- Disable the second-level cache -->
-<!--        <property name="cache.provider_class">org.hibernate.cache.NoCacheProvider</property>-->
-
-        <!-- Echo all executed SQL to stdout -->
-        <property name="show_sql">true</property>
-
-
-        <!-- Mapping files -->
-        <mapping class="org.nomarchia.movieland.entity.Movie"/>
-        <mapping class="org.nomarchia.movieland.entity.Genre"/>
-
-    </session-factory>
-
-</hibernate-configuration>
