+++ src/main/java/org/nomarchia/movieland/MovielandApplicationContext.java
@@ -1,11 +1,35 @@
 package org.nomarchia.movieland;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.orm.jpa.LocalEntityManagerFactoryBean;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+import javax.persistence.EntityManagerFactory;
 
 @Configuration
-@PropertySource("classpath:application.properties")
+@PropertySource(value = {"classpath:application.properties", "classpath:hibernate.properties"})
 @ComponentScan("org.nomarchia.movieland")
+@EnableJpaRepositories(basePackages = {"org.nomarchia.movieland.repository"})
+@EnableTransactionManagement
 public class MovielandApplicationContext {
+    @Bean
+    public LocalEntityManagerFactoryBean entityManagerFactory() {
+        LocalEntityManagerFactoryBean entityManagerFactoryBean = new LocalEntityManagerFactoryBean();
+        entityManagerFactoryBean.setPersistenceUnitName("MovielandDb");
+
+        return entityManagerFactoryBean;
+    }
+
+    @Bean
+    public JpaTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
+        JpaTransactionManager transactionManager = new JpaTransactionManager();
+        transactionManager.setEntityManagerFactory(entityManagerFactory);
+
+        return transactionManager;
+    }
 }
