+++ src/main/java/org/nomarchia/movieland/config/HibernateUtil.java
@@ -4,25 +4,21 @@ import lombok.extern.slf4j.Slf4j;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.cfg.AnnotationConfiguration;
 import org.hibernate.cfg.Configuration;
 import org.nomarchia.movieland.entity.Genre;
 import org.nomarchia.movieland.entity.Movie;
 
-import java.io.File;
-
 @Slf4j
 public class HibernateUtil {
-
     private static final SessionFactory sessionFactory;
 
     static {
         try {
             sessionFactory = new Configuration()
-//                    .addPackage("org.nomarchia.movieland.entity")
-//                    .addClass(Movie.class)
-//                    .addClass(Genre.class)
-                    .configure(new File("hibernate.cfg.xml")).buildSessionFactory();
+                    .addPackage("org.nomarchia.movieland.entity")
+                    .addAnnotatedClass(Movie.class)
+                    .addAnnotatedClass(Genre.class)
+                    .buildSessionFactory();
         } catch (HibernateException e) {
             log.debug("Problem occurred during SessionFactory initialization, exception thrown: ", e);
             throw new ExceptionInInitializerError(e);
@@ -30,6 +26,7 @@ public class HibernateUtil {
     }
 
     public static Session getSession() {
+        log.info("new session is to be opened in HibernateUtil");
         return sessionFactory.openSession();
     }
 }
