+++ src/main/resources/META-INF/persistence.xml
@@ -6,19 +6,7 @@
              version="2.1">
 
     <persistence-unit name="MovielandDb">
-        <properties>
-            <property name="javax.persistence.jdbc.url" value="jdbc:postgresql://localhost:5432/movieland-db" />
-            <property name="javax.persistence.jdbc.user" value="postgres" />
-            <property name="javax.persistence.jdbc.password" value="root" />
-            <property name="javax.persistence.jdbc.driver" value="org.postgresql.Driver" />
 
-            <property name="hibernate.show_sql" value="true" />
-            <property name="hibernate.format_sql" value="true" />
-
-            <property name="hibernate.cache.use_second_level_cache" value="true"/>
-            <property name="hibernate.cache.region.factory_class"
-                      value="org.hibernate.cache.ehcache.EhCacheRegionFactory"/>
-        </properties>
     </persistence-unit>
 
 </persistence>
