+++ src/main/resources/hibernate.properties
@@ -1,0 +1,15 @@
+#Hibernate DB configuration properties
+hibernate.connection.driver_class=org.postgresql.Driver
+hibernate.connection.url=jdbc:postgresql://localhost:5432/movieland-db
+hibernate.connection.username=postgres
+hibernate.connection.password=root
+hibernate.connection.pool_size=1
+hibernate.dialect=org.hibernate.dialect.PostgreSQL9Dialect
+
+#current_session_context_class=thread
+hibernate.show_sql=true
+hibernate.format_sql=true
+
+#cache properties
+hibernate.cache.use_second_level_cache=true
+hibernate.cache.region.factory_class=org.hibernate.cache.ehcache.EhCacheRegionFactory
